+++ tdrs-backend/tdpservice/parsers/schema_defs/utils.py
@@ -1,3 +1,5 @@
+"""Utility functions for parsing."""
+
 from .. import schema_defs
 from tdpservice.data_files.models import DataFile
 import logging
