+++ tdrs-backend/tdpservice/parsers/test/test_validators.py
@@ -6,7 +6,6 @@ from .. import validators
 from .. import schema_defs, util
 from tdpservice.parsers.test.factories import TanfT1Factory, TanfT2Factory, TanfT3Factory, TanfT5Factory, TanfT6Factory
 from tdpservice.parsers.test.factories import SSPM5Factory
-from ..models import ParserError
 
 logger = logging.getLogger(__name__)
 
@@ -1239,6 +1238,7 @@ class TestCat4Validator:
     """Test category four validators."""
 
     def parse_header(self, datafile):
+        """Parse datafile header into header object."""
         rawfile = datafile.file
 
         # parse header, trailer
@@ -1251,6 +1251,7 @@ class TestCat4Validator:
 
     @pytest.fixture
     def s1_records(self):
+        """Return group of TANF Section 1 records."""
         t1 = TanfT1Factory.create()
         t2 = TanfT2Factory.create()
         t3 = TanfT3Factory.create()
@@ -1261,42 +1262,42 @@ class TestCat4Validator:
     def small_correct_file(self, stt_user, stt):
         """Fixture for small_correct_file."""
         return util.create_test_datafile('small_correct_file.txt', stt_user, stt)
-    
+
     @pytest.fixture
     def small_correct_file_header(self, small_correct_file):
         """Return a valid header record."""
-
         header, header_is_valid, header_errors = self.parse_header(small_correct_file)
-        
+
         if not header_is_valid:
             logger.error('Header is not valid: %s', header_errors)
             return None
         return header
-    
+
     @pytest.mark.django_db
     def test_section1_fail(self, small_correct_file_header, small_correct_file, s1_records):
-        cat_four_validator = validators.CatFourValidator(small_correct_file_header, 
+        """Test TANF Section 1 records RPT_MONTH_YEAR don't align with header year and quarter."""
+        cat_four_validator = validators.CatFourValidator(small_correct_file_header,
                                                          util.make_generate_parser_error(small_correct_file, -1))
-        
+
         for record in s1_records:
             record.data_file = small_correct_file
             cat_four_validator.add_record(record, False)
-        
+
         num_errors = cat_four_validator.validate()
 
         assert 4 == num_errors
-    
+
     @pytest.mark.django_db
     def test_section1_pass(self, small_correct_file_header, small_correct_file, s1_records):
-        cat_four_validator = validators.CatFourValidator(small_correct_file_header, 
+        """Test TANF Section 1 records RPT_MONTH_YEAR do align with header year and quarter."""
+        cat_four_validator = validators.CatFourValidator(small_correct_file_header,
                                                          util.make_generate_parser_error(small_correct_file, -1))
-        
+
         for record in s1_records:
             record.RPT_MONTH_YEAR = 202010
             record.data_file = small_correct_file
             cat_four_validator.add_record(record, False)
-        
+
         num_errors = cat_four_validator.validate()
 
         assert 0 == num_errors
-
